requirements.txt
streamlit>=1.28.0
pandas>=2.0.0
openpyxl>=3.1.0
openai>=1.0.0
python-dotenv>=1.0.0
numpy>=1.24.0
langgraph>=0.2.0
langchain>=0.3.0
langchain-core>=0.3.0
langchain-openai>=0.2.0
pydantic>=2.0.0
pydantic-settings>=2.0.0

# Development dependencies
pytest>=7.4.3
pytest-cov>=4.1.0
mypy>=1.9.0
black>=24.3.0
isort>=5.13.2
ruff>=0.3.5
pre-commit>=3.7.0